Adds ConfigMixed.

+++ org/minnen/retiretool/util/Library.java
@@ -282,11 +282,11 @@ public final class Library
     return true;
   }
 
-  public static long numBits(long x)
+  public static int numBits(long x)
   {
     x = x - ((x >>> 1) & 0x55555555);
     x = (x & 0x33333333) + ((x >>> 2) & 0x33333333);
-    return (((x + (x >>> 4)) & 0x0F0F0F0F) * 0x01010101) >>> 24;
+    return (int) ((((x + (x >>> 4)) & 0x0F0F0F0F) * 0x01010101) >>> 24);
   }
 
   public static int numBits(int x)
